@@ -3,7 +3,7 @@
   <PropertyGroup>
 	<TargetFrameworks Condition="'$(LibraryFrameworks)'==''">net3.5;netstandard2.0;netstandard2.1;net6.0;net7.0</TargetFrameworks>
 	<TargetFrameworks Condition="'$(LibraryFrameworks)'!=''">$(LibraryFrameworks)</TargetFrameworks>
-	<LangVersion>9.0</LangVersion>
+	<LangVersion>10.0</LangVersion>
 	<Nullable>enable</Nullable>
 	<GenerateDocumentationFile>true</GenerateDocumentationFile>  
 	<NeutralLanguage>en-US</NeutralLanguage>
